+++ Tests/ToolBox/RandomDataGenerator/SymbolGeneratorTests.cs
@@ -13,9 +13,59 @@ namespace QuantConnect.Tests.ToolBox.RandomDataGenerator
         private const int Seed = 123456789;
         private static readonly IRandomValueGenerator _randomValueGenerator = new RandomValueGenerator(Seed);
 
+
+        [Test]
+        [TestCase(2, 5)]
+        [TestCase(3, 3)]
+        [TestCase(1, 4)]
+        public void NextUpperCaseString_CreatesString_WithinSpecifiedMinMaxLength(int min, int max)
+        {
+            var symbolGenerator = new Mock<SymbolGenerator>(Mock.Of<RandomDataGeneratorSettings>(), new RandomValueGenerator()).Object;
+            var str = symbolGenerator.NextUpperCaseString(min, max);
+            Assert.LessOrEqual(min, str.Length);
+            Assert.GreaterOrEqual(max, str.Length);
+        }
+
+        [Test]
+        public void NextUpperCaseString_CreatesUpperCaseString()
+        {
+            var symbolGenerator = new Mock<SymbolGenerator>(Mock.Of<RandomDataGeneratorSettings>(), new RandomValueGenerator()).Object;
+            var str = symbolGenerator.NextUpperCaseString(10, 10);
+            Assert.IsTrue(str.All(char.IsUpper));
+        }
+
+        [Test]
+        [TestCase(SecurityType.Option)]
+        [TestCase(SecurityType.Future)]
+        public void ThrowsArgumentException_ForDerivativeSymbols(SecurityType securityType)
+        {
+            var symbolGenerator = new Mock<SymbolGenerator>(Mock.Of<RandomDataGeneratorSettings>(), new RandomValueGenerator()).Object;
+            Assert.Throws<ArgumentException>(() =>
+                symbolGenerator.NextSymbol(securityType, Market.USA)
+            );
+        }
+
         [TestFixture]
         public class SpotSymbolGeneratorTests
         {
+            private SymbolGenerator _symbolGenerator;
+            private DateTime _minExpiry = new(2000, 01, 01);
+            private DateTime _maxExpiry = new(2001, 01, 01);
+
+            [SetUp]
+            public void Setup()
+            {
+                // initialize using a seed for deterministic tests
+                _symbolGenerator = new FutureSymbolGenerator(
+                    new RandomDataGeneratorSettings()
+                    {
+                        Market = Market.CME,
+                        Start = _minExpiry,
+                        End = _maxExpiry
+                    },
+                    _randomValueGenerator);
+            }
+
             [Test]
             [TestCase(SecurityType.Equity, Market.USA, true)]
             [TestCase(SecurityType.Cfd, Market.FXCM, false)]
@@ -38,6 +88,87 @@ namespace QuantConnect.Tests.ToolBox.RandomDataGenerator
 
                 Assert.AreEqual(expected, symbolGenerator.GetAvailableSymbolCount());
             }
+
+            [Test]
+            [TestCase(SecurityType.Equity, Market.USA)]
+            [TestCase(SecurityType.Cfd, Market.FXCM)]
+            [TestCase(SecurityType.Cfd, Market.Oanda)]
+            [TestCase(SecurityType.Forex, Market.FXCM)]
+            [TestCase(SecurityType.Forex, Market.Oanda)]
+            [TestCase(SecurityType.Crypto, Market.GDAX)]
+            [TestCase(SecurityType.Crypto, Market.Bitfinex)]
+            public void NextSymbol_CreatesSymbol_WithRequestedSecurityTypeAndMarket(SecurityType securityType, string market)
+            {
+                var symbolGenerator = new SpotSymbolGenerator(new RandomDataGeneratorSettings
+                {
+                    SecurityType = securityType,
+                    Market = market
+                }, _randomValueGenerator);
+
+                var symbol = symbolGenerator.GenerateSingle();
+
+                Assert.AreEqual(securityType, symbol.SecurityType);
+                Assert.AreEqual(market, symbol.ID.Market);
+            }
+
+            [Test]
+            [TestCase(SecurityType.Equity, Market.USA)]
+            [TestCase(SecurityType.Cfd, Market.FXCM)]
+            [TestCase(SecurityType.Cfd, Market.Oanda)]
+            [TestCase(SecurityType.Forex, Market.FXCM)]
+            [TestCase(SecurityType.Forex, Market.Oanda)]
+            [TestCase(SecurityType.Crypto, Market.GDAX)]
+            [TestCase(SecurityType.Crypto, Market.Bitfinex)]
+            public void NextSymbol_CreatesSymbol_WithEntryInSymbolPropertiesDatabase(SecurityType securityType, string market)
+            {
+                var symbolGenerator = new SpotSymbolGenerator(new RandomDataGeneratorSettings
+                {
+                    SecurityType = securityType,
+                    Market = market
+                }, _randomValueGenerator);
+
+                var symbol = symbolGenerator.GenerateSingle();
+
+                var db = SymbolPropertiesDatabase.FromDataFolder();
+                if (db.ContainsKey(market, SecurityDatabaseKey.Wildcard, securityType))
+                {
+                    // there is a wildcard entry, so no need to check whether there is a specific entry for the symbol
+                    Assert.Pass();
+                }
+                else
+                {
+                    // there is no wildcard entry, so there should be a specific entry for the symbol instead
+                    Assert.IsTrue(db.ContainsKey(market, symbol, securityType));
+                }
+            }
+
+            [Test]
+            [TestCase(SecurityType.Cfd, Market.FXCM)]
+            [TestCase(SecurityType.Cfd, Market.Oanda)]
+            [TestCase(SecurityType.Forex, Market.FXCM)]
+            [TestCase(SecurityType.Forex, Market.Oanda)]
+            [TestCase(SecurityType.Crypto, Market.GDAX)]
+            [TestCase(SecurityType.Crypto, Market.Bitfinex)]
+            public void NextSymbol_ThrowsNoTickersAvailableException_WhenAllSymbolsGenerated(SecurityType securityType, string market)
+            {
+                var db = SymbolPropertiesDatabase.FromDataFolder();
+                var symbolCount = db.GetSymbolPropertiesList(market, securityType).Count();
+
+                var symbolGenerator = new SpotSymbolGenerator(new RandomDataGeneratorSettings
+                {
+                    SecurityType = securityType,
+                    Market = market
+                }, _randomValueGenerator);
+
+                for (var i = 0; i < symbolCount; i++)
+                {
+                    symbolGenerator.GenerateSingle();
+                }
+
+                Assert.Throws<NoTickersAvailableException>(() =>
+                    symbolGenerator.GenerateSingle()
+                );
+            }
         }
 
         [TestFixture]
@@ -58,7 +189,7 @@ namespace QuantConnect.Tests.ToolBox.RandomDataGenerator
                         Start = _minExpiry,
                         End = _maxExpiry
                     },
-                    new RandomValueGenerator(Seed));
+                    _randomValueGenerator);
             }
 
             [Test]
